load-balancer/requirements.txt
Flask==2.3.3
gunicorn==21.2.0
eventlet==0.33.3
requests==2.31.0
redis==5.0.1
flask-cors==4.0.0
flask-socketio==5.3.6
prometheus-client==0.18.0
python-dotenv==1.0.0
pyyaml==6.0.1